+++ tests/functional/test_k.py
@@ -36,7 +36,7 @@ class Test_k(unittest.PatchFeedCommandXpos, unittest.FunctionalTestCase):
         self.eq('aaa\naaa bb|b\n', 'n_k', 'aa|a\naaa bbb\n')
         self.eq('aaa\naaa bbb |ccc\n', 'n_k', 'aa|a\naaa bbb ccc\n')
 
-    def test_n_tabs(self):
+    def test_n_no_tabs_to_spaces(self):
         self.settings().set('translate_tabs_to_spaces', False)
         self.eq('\t\taaa\naaa bbb |ccc\n', 'n_k', '\t\t|aaa\naaa bbb ccc\n')
 
@@ -61,14 +61,52 @@ class Test_k(unittest.PatchFeedCommandXpos, unittest.FunctionalTestCase):
         self.eq('r_1x\nfizz\n|buzz\n|4x', 'l_k', 'r_1x\n|fizz\nbuzz\n|4x')
         self.eq('1x\n|fizz\nbuzz\n|4x', 'l_k', '1x\n|fizz\n|buzz\n4x')
 
-    def test_b(self):
-        self.eq('fizz buzz\nfiz|z bu|zz\n', 'b_k', 'fiz|z bu|zz\nfiz|z bu|zz\n')
-        self.eq('fizz buzz\nfiz|z bu|zz\nfiz|z bu|zz\n', 'b_k', 'fizz buzz\nfiz|z bu|zz\nfizz buzz\n')
-        self.eq('fiz|z bu|zz\nfiz|z bu|zz\n', 'b_k', 'fiz|z bu|zz\nfizz buzz\n')
-        self.eq('fiz|z bu|zz\nfiz|z bu|zz\nfiz|z bu|zz\n', 'b_k', 'fiz|z bu|zz\nfiz|z bu|zz\nfizz buzz\n')
-        self.eq('r_fi|zz bu|zz\n', 'b_k', 'r_fi|zz bu|zz\n')
-        self.eq('fizzbuzz\nfizzbuzz\nfizzbuzz\nfi|zzbu|zz\n', 'b_2k', 'fizzbuzz\nfi|zzbu|zz\nfi|zzbu|zz\nfi|zzbu|zz\n')
-        self.eq('fizz\nfizz\nfizz\nfizz\nf|iz|z\n', 'b_3k', 'fizz\nf|iz|z\nf|iz|z\nf|iz|z\nf|iz|z\n')
-
     def test_d(self):
         self.eq('1\n2\n|3\n4\n5', 'dk', '1\n|4\n5')
+
+    def test_b(self):
+        self.eq('fizz\n|fizz|\n|fizz|\n|fizz|\nfizz', 'b_k', 'fizz\n|fizz|\n|fizz|\nfizz\nfizz')
+        self.eq('fizz\n|fizz|\n|fizz|\n|fizz|\n|fizz|\nfizz', 'b_3k', 'fizz\n|fizz|\nfizz\nfizz\nfizz\nfizz')
+        self.eq('fizz\n|fizz|\n|fizz|\nfizz\nfizz', 'b_k', 'fizz\n|fizz|\nfizz\nfizz\nfizz')
+        self.eq('u_fizz\n|fi|zz', 'b_k', 'u_|fi|zz\n|fi|zz')
+        self.eq('r_fizz\n|fizz|\n|fizz|\n|fizz|\nfizz', 'b_k', 'r_fizz\n|fizz|\n|fizz|\nfizz\nfizz')
+        self.eq('fizz\n|fizz|\n|\n||fizz|\nfizz', 'b_k', 'fizz\n|f|izz\n|\n|fizz\nfizz')
+        self.eq('fizz\n\nfi|zz|\nfizz\n', 'b_k', 'r_u_fizz\n|\n||fiz|z\nfizz\n')
+        self.eq('fizz\nf\nfi|zz|\nfizz\n', 'b_k', 'r_u_fizz\nf|\n|f|iz|z\nfizz\n')
+        self.eq('fizz\nfi\nfi|zz|\nfizz\n', 'b_k', 'u_fizz\nfi|\n|fi|z|z\nfizz\n')
+        self.eq('fizz\nfiz\nfi|zz|\nfizz\n', 'b_k', 'u_fizz\nfi|z\n|fi|zz|\nfizz\n')
+        self.eq('fizz\nfizz\nfi|zz|\nfizz\n', 'b_k', 'u_fizz\nfi|zz|\nfi|zz|\nfizz\n')
+        self.eq('r_fizz\n\nfi|zz|\nfizz\n', 'b_k', 'r_u_fizz\n|\n||fizz|\nfizz\n')
+        self.eq('r_fizz\nf\nfi|zz|\nfizz\n', 'b_k', 'r_u_fizz\nf|\n|f|izz|\nfizz\n')
+        self.eq('r_fizz\nfi\nfi|zz|\nfizz\n', 'b_k', 'r_u_fizz\nfi|\n|fi|zz|\nfizz\n')
+        self.eq('r_fizz\nfiz\nfi|zz|\nfizz\n', 'b_k', 'r_u_fizz\nfi|z\n|fi|zz|\nfizz\n')
+        self.eq('r_fizz\nfizz\nfi|zz|\nfizz\n', 'b_k', 'r_u_fizz\nfi|zz|\nfi|zz|\nfizz\n')
+        self.eq('u_fizz\nfizz\nfi|zz|\nfizz\n', 'b_k', 'u_fizz\nfi|zz|\nfi|zz|\nfizz\n')
+        self.eq('d_fizz\nfizz\nfi|zz|\nfizz\n', 'b_k', 'u_fizz\nfi|zz|\nfi|zz|\nfizz\n')
+        self.eq('d_fizz\nfi|zz|\nfi|zz|\nfizz\n', 'b_k', 'd_fizz\nfi|zz|\nfizz\nfizz\n')
+        self.eq('d_fi|zz|\nfi|zz|\nfi|zz|\nfizz\n', 'b_k', 'd_fi|zz|\nfi|zz|\nfizz\nfizz\n')
+        self.eq('r_u_fizz\nfizz\nfi|zz|\nfizz\n', 'b_k', 'r_u_fizz\nfi|zz|\nfi|zz|\nfizz\n')
+        self.eq('r_d_fizz\nfizz\nfi|zz|\nfizz\n', 'b_k', 'r_u_fizz\nfi|zz|\nfi|zz|\nfizz\n')
+        self.eq('d_fizz |buzz|\nf\nfizz |buzz|\n', 'b_k', 'r_d_f|izz b|uzz\nf|\n|fizz buzz\n')
+        self.eq('d_fizz\nfizz\nf|iz|z\nf|iz|z\n', 'b_2k', 'u_fizz\nf|iz|z\nf|iz|z\nfizz\n')
+        self.eq('r_d_fizz\nfizz\nf|iz|z\nf|iz|z\n', 'b_2k', 'r_u_fizz\nf|iz|z\nf|iz|z\nfizz\n')
+        self.eq('d_fizz\nfizz\nfizz\nf|iz|z\nf|iz|z\n', 'b_3k', 'u_fizz\nf|iz|z\nf|iz|z\nf|iz|z\nfizz\n')
+        self.eq('d_fizz\nfizz\n\nf|iz|z\nf|iz|z\n', 'b_3k', 'u_fizz\nf|iz|z\n\nf|iz|z\nfizz\n')
+        self.eq('d_fizz\nfizz\nf\nf|iz|z\nf|iz|z\n', 'b_3k', 'u_fizz\nf|iz|z\nf|\n|f|iz|z\nfizz\n')
+        self.eq('d_fizz\nfizz\nfi\nf|iz|z\nf|iz|z\n', 'b_3k', 'u_fizz\nf|iz|z\nf|i\n|f|iz|z\nfizz\n')
+        self.eq('d_fizz\n\nfi\nf|iz|z\nf|iz|z\n', 'b_3k', 'r_u_fizz\n|\n||fi|\n|fi|zz\nfizz\n')
+        self.eq('d_fizz\nx\nfizz |buz|z\nfizz |buz|z\n', 'b_3k', 'r_u_fizz|\n|x\nfizz| b|uzz\nfizz buzz\n')
+
+        self.setXpos(5)
+        self.eq('r_|fizz buz|z\n|\n|\n', 'b_k', 'r_fizz |buz|z\n\n\n')
+        self.eq('|fi|zz buzz\n|f\n|\n', 'b_k', '|fizz b|uzz\nf\n\n')
+        self.eq('u_fizz buzz\nf\nfizz |buzz|\n', 'b_2k', 'u_fizz |buzz|\nf\nfizz |buzz|\n')
+        self.eq('u_fizz\nf\nfizz |buzz|\n', 'b_2k', 'r_u_fizz|\n|f\nfizz| b|uzz\n')
+        self.eq('u_fizz \nf\nfizz |buzz|\n', 'b_2k', 'u_fizz |\n|f\nfizz |b|uzz\n')
+        self.eq('fizz buzz\nf\nfizz |buz|z\n', 'b_2k', 'u_fizz |buz|z\nf\nfizz |buz|z\n')
+        self.eq('r_u_fizz \nf\nfizz |buzz|\n', 'b_2k', 'r_u_fizz |\n|f\nfizz |buzz|\n')
+        self.eq('r_d_fizz |buzz|\nf\nfizz |buzz|\nfizz |buzz|\n', 'b_k', 'r_d_fizz |buzz|\nf\nfizz |buzz|\nfizz buzz\n')  # noqa: E501
+        self.eq('r_d_f|izz b|uzz\nf|\n|fizz buzz\n', 'b_k', 'd_fizz |b|uzz\nf\nfizz buzz\n')
+
+        self.setXpos(7)
+        self.eq('r_u_fizz buzz\nf|\n|f|izz b|uzz\n', 'b_5k', 'u_fizz |buz|z\nf\nfizz |buz|z\n')